requirement.txt
pillow
pytube